requirements.txt
# Core dependencies
pymupdf==1.24.0
pandas==2.2.0
python-dotenv==1.0.0
pyarrow>=14.0.0

# NLP and Embeddings
sentence-transformers==2.7.0
langchain==0.2.0
langchain-community==0.2.0

# Graph Database
neo4j==5.20.0

# Vector Store
chromadb==0.5.0

# LLM Integration
ollama==0.2.0

# Retrieval Enhancements
rank-bm25==0.2.2

# Web Interface
streamlit==1.32.0

# Utilities
tqdm==4.66.0
pyyaml==6.0.1
numpy>=1.24.0